+++ packages/awino-ui/src/app/state/slices/account.ts
@@ -2,11 +2,11 @@ import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 interface AccountState {
   connected: boolean;
-  activeAccount: string | null;
+  walletAddress: string | null;
 }
 
 const initialState: AccountState = {
-  activeAccount: null,
+  walletAddress: null,
   connected: false,
 };
 
@@ -14,22 +14,13 @@ export const accountSlice = createSlice({
   name: 'account',
   initialState,
   reducers: {
-    setActiveAccount: (state, action: PayloadAction<string | undefined | null>) => {
-      state.activeAccount = action.payload === null ? null : action.payload;
+    setAccount: (state, action: PayloadAction<string | undefined | null>) => {
+      state.walletAddress = action.payload === null ? null : action.payload;
       state.connected = !!action.payload;
-    },
-    connect: (state, action?: PayloadAction<string>) => {
-      /* TODO WIP Check if account is valid */
-      state.activeAccount = action.payload || 'ACTIVE_ACCOUNT';
-      state.connected = true;
-    },
-    disconnect: (state) => {
-      state.activeAccount = null;
-      state.connected = false;
     },
   },
 });
 
-export const { setActiveAccount, connect, disconnect } = accountSlice.actions;
+export const { setAccount } = accountSlice.actions;
 
 export default accountSlice.reducer;