+++ packages/awino-ui/src/pages/portfolio.tsx
@@ -24,6 +24,7 @@ export const getServerSideProps = storeWrapper.getServerSideProps((store) => asy
 
   return {
     props: {
+      protected: true,
       ...(await serverSideTranslations(locale, ['common', ns])),
     },
   };